\name{moesim}
\alias{moesim}
\docType{data}
\title{
Simulated data
}
\description{
A simulated two-component mixutre of gaussians.
}
\usage{data(moesim)}
\format{
 $ x: the matrix of covariates
 $ y:  the vector of response variable
}
\source{Simulated by Abbas Khalili.}
\examples{data(moesim)}
\keyword{datasets}
